+++ NEWS.md
@@ -1,5 +1,11 @@
 # Spectra 1.1
 
+## Changes in 1.1.5
+
+- Add `[[,Msbackend` and `[[<-,MsBackend` methods (issue
+  [149](https://github.com/rformassspectrometry/Spectra/issues/149)).
+- Add `[[,Spectra` and `[[<-,Spectra` methods.
+
 ## Changes in 1.1.4
 
 - Fix issue with `labelCol` in `plotSpectra` (issue